Incorporate Mirror Protobuf
Reworked approach for incroporating both gossip and
mirror node protobuf into single hapi-proto package
that accounts for namespace collisions in the protobuf.

Where there are mirror specific objects that have names
that collide with services probuf names, the names are
now aliased with the prefix 'Mirror' to distinguish them
from gossip node versions which retain their original
names.  This should be backwards compatible.

This results in the end of hapi-mirror-protobuf package.