Conversion to Monorepo
Converted project to monorepo to support
the following packages:

hapi-proto
hapi-util
hapi-mirror

All of which are a work in progress.  hapi-util
and hapi-mirror are projects still in discovery
and may drastically change over time.

hapi-proto is a one-to-one mapping of the offical
hedera protobuf project, it will change over time at
the rate that
https://github.com/hashgraph/hedera-protobufs
project changes.